+++ lib/ui/page/home/page/user/view.dart
@@ -141,7 +141,7 @@ class UserView extends StatelessWidget {
   /// Returns the [User.avatar] visual representation.
   Widget _avatar(UserController c, BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.fromLTRB(8, 0, 8, 0),
+      padding: const EdgeInsets.fromLTRB(8, 0, 8, 4),
       child: Center(
         child: ConstrainedBox(
           constraints: context.isNarrow
@@ -336,7 +336,8 @@ class UserView extends StatelessWidget {
     return Row(
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
-        StyledBackButton(),
+        const SizedBox(width: 4),
+        const StyledBackButton(),
         const SizedBox(width: 8),
         Expanded(
           child: Column(